Annotation Processor disabled.
If we generate ExtensionModule resource file using annotation processor,
Intellij Idea IDE is not recognizing this file and it is not showing the
extension methods in IDE code hints.

Also, the IDE is unable to find annotation processor if it is available
in the source itself. i.e., only if we add jar as dependency for the
lsdsl-apt project, only then IDE runs the annoation processor. Even
thought it is running, it is not recognizing it.

So, I have brought back the static resource file, I will spend some more
time in this and bring this feature